# apps/api/app/api/v1/analytics.py
"""
EchoPress AI Backend - Analytics API
Endpoints for content performance analytics
"""

from fastapi import APIRouter, Depends, HTTPException, status, Query
from typing import List, Optional

from app.core.auth import get_current_user
from app.schemas.analytics import (
    AnalyticsResponse,
    AnalyticsListResponse,
    ContentPerformanceResponse,
    PerformanceListResponse,
    AnalyticsSummary
)
from app.schemas.user import User
from app.services.analytics import AnalyticsService

router = APIRouter()

@router.get("/episodes/{episode_id}", response_model=AnalyticsListResponse)
async def get_episode_analytics(
    episode_id: str,
    page: int = Query(1, ge=1, description="Page number"),
    limit: int = Query(20, ge=1, le=100, description="Items per page"),
    current_user: User = Depends(get_current_user)
):
    """
    Get analytics for an episode
    """
    try:
        analytics_service = AnalyticsService()
        # TODO: Implement episode analytics retrieval logic
        raise HTTPException(
            status_code=status.HTTP_501_NOT_IMPLEMENTED,
            detail="Episode analytics retrieval not yet implemented"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to get episode analytics: {str(e)}"
        )

@router.get("/drafts/{draft_id}", response_model=PerformanceListResponse)
async def get_draft_performance(
    draft_id: str,
    page: int = Query(1, ge=1, description="Page number"),
    limit: int = Query(20, ge=1, le=100, description="Items per page"),
    current_user: User = Depends(get_current_user)
):
    """
    Get performance metrics for a draft
    """
    try:
        analytics_service = AnalyticsService()
        # TODO: Implement draft performance retrieval logic
        raise HTTPException(
            status_code=status.HTTP_501_NOT_IMPLEMENTED,
            detail="Draft performance retrieval not yet implemented"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to get draft performance: {str(e)}"
        )

@router.get("/overview", response_model=AnalyticsSummary)
async def get_analytics_overview(
    workspace_id: Optional[str] = Query(None, description="Workspace ID"),
    date_range: Optional[str] = Query("30d", description="Date range"),
    current_user: User = Depends(get_current_user)
):
    """
    Get analytics overview for dashboard
    """
    try:
        analytics_service = AnalyticsService()
        # TODO: Implement analytics overview logic
        raise HTTPException(
            status_code=status.HTTP_501_NOT_IMPLEMENTED,
            detail="Analytics overview not yet implemented"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to get analytics overview: {str(e)}"
        )

@router.get("/performance")
async def get_performance_metrics(
    workspace_id: Optional[str] = Query(None, description="Workspace ID"),
    metric: Optional[str] = Query(None, description="Metric type"),
    period: Optional[str] = Query("30d", description="Time period"),
    current_user: User = Depends(get_current_user)
):
    """
    Get performance metrics over time
    """
    try:
        analytics_service = AnalyticsService()
        # TODO: Implement performance metrics logic
        raise HTTPException(
            status_code=status.HTTP_501_NOT_IMPLEMENTED,
            detail="Performance metrics not yet implemented"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to get performance metrics: {str(e)}"
        )

@router.get("/engagement")
async def get_engagement_metrics(
    workspace_id: Optional[str] = Query(None, description="Workspace ID"),
    current_user: User = Depends(get_current_user)
):
    """
    Get engagement metrics
    """
    try:
        analytics_service = AnalyticsService()
        # TODO: Implement engagement metrics logic
        raise HTTPException(
            status_code=status.HTTP_501_NOT_IMPLEMENTED,
            detail="Engagement metrics not yet implemented"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to get engagement metrics: {str(e)}"
        )