# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: EchoPress AI API
  description: AI-powered podcast to blog post converter with transcription, RAG, and SEO optimization
  version: 1.0.0
  contact:
    name: EchoPress AI Support
    email: support@echopress.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.echopress.ai
    description: Production server
paths:
...
  /api/v1/episodes/{episode_id}:
    get:
      summary: Get episode
      description: Retrieve a specific episode by ID
      tags:
        - Episodes
      parameters:
        - name: episode_id
          in: path
          required: true
          description: Episode ID
          schema:
            type: string
      responses:
        '200':
          description: Episode retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodeResponse'
        '404':
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Episode:
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "The Future of AI in Podcasting"
        description:
          type: string
          example: "Exploring how artificial intelligence is transforming the podcast industry"
        audioUrl:
          type: string
          format: uri
          example: "https://example.com/episode1.mp3"
        duration:
          type: integer
          description: Duration in seconds
          example: 3600
        transcript:
          type: string
          example: "Welcome to our podcast about AI in podcasting..."
        status:
          type: string
          enum: [pending, processing, completed, failed]
          example: "completed"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T11:30:00Z"
      required:
        - id
        - title
        - description
        - audioUrl
        - duration
        - status
        - createdAt
        - updatedAt
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
      required:
        - success
        - message
...
    EpisodeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Episode'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
      required:
        - success
        - message
...